pyproject.toml
[project]
name = "omnimisep1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "google-api-python-client>=2.192.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "langchain-core>=1.2.18",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
]
